+++ src/pat/colour-picker.js
@@ -1,0 +1,23 @@
+/**
+ * Patterns pat-polyfill-colour - Polyfill for colour inputs.
+ *
+ * Copyright 2014 Marko Durkovic
+ * Copyright 2014 Simplon B.V. - Wichert Akkerman
+ */
+define([
+    "pat-registry",
+    "spectrum"
+], function(registry) {
+    var _ = {
+        name: "polyfill-color",
+        trigger: "input.pat-colourpicker,input.pat-colorpicker",
+        init: function($el) {
+            return $el.spectrum();
+        }
+    };
+
+    registry.register(_);
+    return _;
+});
+
+// vim: sw=4 expandtab